Split Edit Person actioner workbench to multiple files
