+++ app/webpacker/components/Tickets/TicketWorkbenches.jsx
@@ -1,8 +1,8 @@
 import { ticketTypes, ticketStakeholderRoles } from '../../lib/wca-data.js.erb';
-import EditPersonTicketWorkbench from './TicketWorkbenches/EditPersonTicketWorkbench';
+import EditPersonActionerView from './TicketWorkbenches/EditPersonActionerView';
 
 export default {
   [ticketTypes.edit_person]: {
-    [ticketStakeholderRoles.actioner]: EditPersonTicketWorkbench,
+    [ticketStakeholderRoles.actioner]: EditPersonActionerView,
   },
 };
